move trimTrailingSlash to remote providers
